(maint) Remove metadata.json template from module skeleton
Before this patch the generate action included an empty metadata.json
file inside of the generated puppet module. This file is unnecessary as
it is automatically generated at build time.

Another issue with generating an empty metadata.json file is that it
gets picked up during the build process and included in the list of
checksummed files. The metadata file is then overwritten by the build
process which means the checksum will not match after install.

These side affects lead to the changes and uninstall actions to always
report that the module has local changes.

This patch fixes this issue by removing the metadata.json from the
module skeleton template directory.

This patch does not introduce any other changes in behavior.