(maint) Fixing failing search acceptance test.
Without this patch the communication_error search acceptance test does
not pass. This is due to the use of the wrong error message to
communicate the failure and the improper exit code of 0.

This patch fixes this issue by using the right error message as
specified by UX designs and exiting with a status of 1.